extract caching process
